Adding flask-bouncer to extensions
flask-bouncer is a authorization library based on Ryan Bates ruby library CanCan.  I am actively support both flask-bouncer and flask.  Hope it is a useful extension for the flask community.

Feel free to make it an approved extension (but I bet I need a bit more of a track record first ;-)

Thanks guys